+++ vulkano/src/buffer/slice.rs
@@ -17,7 +17,6 @@ use buffer::traits::BufferInner;
 use buffer::traits::TypedBufferAccess;
 use device::Device;
 use device::DeviceOwned;
-use device::Queue;
 use image::ImageAccess;
 use sync::AccessError;
 
@@ -242,7 +241,7 @@ unsafe impl<T: ?Sized, B> BufferAccess for BufferSlice<T, B>
     }
 
     #[inline]
-    fn try_gpu_lock(&self, exclusive_access: bool, queue: &Queue, range: Range<usize>) -> Result<(), AccessError> {
+    fn try_gpu_lock(&self, exclusive_access: bool, range: Range<usize>) -> Result<(), AccessError> {
         // TODO: subranges. at the time of writing we should never hit this
         // assert because the only place calling this function is
         // SyncCommandBuffer and it's always locking the entire BufferAccess (so
@@ -250,7 +249,7 @@ unsafe impl<T: ?Sized, B> BufferAccess for BufferSlice<T, B>
         assert_eq!(range, 0..self.size());
 
         let offset = self.inner().offset;
-        self.resource.try_gpu_lock(exclusive_access, queue, offset..offset + self.size())
+        self.resource.try_gpu_lock(exclusive_access, offset..offset + self.size())
     }
 
     #[inline]